removed commented string

+++ src/firmware/firmware.ino
@@ -4559,9 +4559,8 @@ void mapWeather() {
     adapted_weather_leds[7] = (weather_leds[25] + weather_leds[7]) / 2.0f;
   }
   for (uint16_t i = 0; i < strip->PixelCount(); i++) {
-    float hue = processWeather(adapted_weather_leds[i]);
+    int hue = processWeather(adapted_weather_leds[i]);
     RGBColor rgb = hue2rgb(hue);
-    // strip->SetPixelColor(i, HslColor(processWeather(adapted_weather_leds[i]), 1.0, settings.current_brightness / 200.0f));
     strip->SetPixelColor(i, RgbColor(rgb.r, rgb.g, rgb.b).Dim(round(settings.current_brightness * 255 / 200.0f)));
   }
   strip->Show();
